update: location_enum, util_enum
