+++ src/algorithm/location.rs
@@ -1,3 +1,5 @@
+pub mod location_enum;
 pub mod point_2_circle_2;
 pub mod point_2_polygon_2;
 pub mod point_2_segment_2;
+pub mod point_2_triangle_2;
