+++ src/kernel.rs
@@ -9,5 +9,6 @@ pub mod polygon_2;
 pub mod ray_2;
 pub mod segment_2;
 pub mod triangle_2;
+pub mod util_enum;
 pub mod vector_2;
 pub mod vertex_2;
